tests passing

+++ tests/PictureTest.php
@@ -6,10 +6,12 @@ use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class PictureTest extends TestCase
 {
-  public $picture = '.\tests\2721.jpg';
-  public $publicPath = '.\tests';
-  public $finalPath = '.\tests' . DIRECTORY_SEPARATOR . 'uploads';
-  public $finalPicturePath = '.\tests' . DIRECTORY_SEPARATOR . 'uploads' . DIRECTORY_SEPARATOR . 'picture.jpg';
+  public $picture = __DIR__ . '/2721.jpg';
+  public $testPath = __DIR__ .'/uploads/pictures/allofthem/';
+
+  public $publicPath = '.\\';
+  public $finalPath = '.\\' . DIRECTORY_SEPARATOR . 'uploads';
+  public $finalPicturePath = '.\\' . DIRECTORY_SEPARATOR . 'uploads' . DIRECTORY_SEPARATOR . 'picture.jpg';
   public $newName = 'picture';
   /**
    * Setup the test environment.
@@ -17,6 +19,7 @@ class PictureTest extends TestCase
   protected function setUp()
   {
     parent::setUp();
+    $this->rrmdir('.tests\uploads');
     // rmdir($this->publicPath);
     // rmdir($this->publicPath . '/uploads');
       // Your code here
@@ -39,40 +42,56 @@ class PictureTest extends TestCase
     ];
   }
 
-  // public function test_creates_directories()
-  // {
-  //   $this->assertDirectoryNotExists($this->publicPath, 'O diretório público nao existe');
-  //   Picture::createUploadDir($this->publicPath, '/uploads');
-  //   $this->assertDirectoryExists($this->publicPath, 'Diretorio publico criado');
-  //   rmdir($this->publicPath, true);
-  // }
+  public function rrmdir($dir) {
+   if (is_dir($dir)) {
+     $objects = scandir($dir);
+     foreach ($objects as $object) {
+       if ($object != "." && $object != "..") {
+         if (is_dir($dir."/".$object))
+           rrmdir($dir."/".$object);
+         else
+           unlink($dir."/".$object);
+       }
+     }
+     rmdir($dir);
+   }
+  }
 
-  // public function test_moves_picture()
-  // {
-  //   $picture = new UploadedFile($this->picture, $this->picture);
+  public function makeTestFile()
+  {
+    return new UploadedFile($this->picture, $this->picture);
+  }
 
-  //   $this->assertFileNotExists($this->finalPicturePath, 'The picture is not in its place');
-  //   Picture::makePictureFromWidth($picture, 300, $this->newName, $this->finalPath, true);
-  //   $this->assertFileExists($this->finalPicturePath, 'the picture exists!');
-  // }
-  // 
   public function test_create_whole_path()
   {
     $testPath = 'tests/uploads/netflix/garbage';
     $test = Picture::preparePath($testPath);
-    $this->assertDirectoryExists('tests/uploads', 'Uploads existe');
-    $this->assertDirectoryExists('tests/uploads/netflix', 'netflix existe');
     $this->assertDirectoryExists($testPath, '$testPath ' . $testPath , ' existe');
-    $this->assertContains($test, './' .$testPath, 'test ' . $test . ' e ' . $testPath);
-    $this->delTree('./tests/uploads');
+    $this->assertEquals('./tests/uploads/netflix/garbage/', $test);
+    // $this->delTree('./tests/uploads');
+  }
+
+  public function test_make_picture()
+  {
+    $options = [
+      'name' => 'picture',
+      'path' => $this->testPath,
+      'action' => 'resize',
+      'width' => 300,
+      'height' => null,
+      'maintainAspectRatio' => true,
+      'quality' => 90,
+    ];
+    Picture::_makePicture($this->makeTestFile(), $options);
+    $this->assertFileExists('uploads/picture/long/path/picture.jpg');
   }
 
 
-  public function delTree($dir) { 
-   $files = array_diff(scandir($dir), array('.','..'));
-    foreach ($files as $file) { 
-      (is_dir("$dir/$file")) ? $this->delTree("$dir/$file") : unlink("$dir/$file"); 
-    } 
-    return rmdir($dir); 
-  } 
+  // public function delTree($dir) {
+  //  $files = array_diff(scandir($dir), array('.','..'));
+  //   foreach ($files as $file) {
+  //     (is_dir("$dir/$file")) ? $this->delTree("$dir/$file") : unlink("$dir/$file");
+  //   }
+  //   return rmdir($dir);
+  // }
 }
